Add typings for manage_data.
